replace utils & update render and remove func
